monolith/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
pyjwt==2.9.0
requests==2.32.3
werkzeug==3.0.3
